Revert "feat: redisigned nft page"
