+++ packages/frontend/src/app/components/NFT/NFTDeal/NFTDeal.tsx
@@ -14,29 +14,12 @@ export type NFTDealProps = PropsWithChildren<{
   reFetchOrder?: () => void // currently order is refreshed only when it's created and cancelled
 }>
 
-const NFTDealStyle = styled('div', {
-  width: '400px',
-  height: '160px',
-  background: 'linear-gradient(0deg, rgba(255, 255, 255, 0.95), rgba(255, 255, 255, 0.95)), #232528',
-  borderRadius: '20px',
-  display: 'flex',
-  justifyContent: 'center',
-  alignItems: 'center',
-  flexDirection: 'column',
-  '@md': {
-    width: '100%',
-    height: '201px'
-  }
-})
-
 const DealContainerInfo = styled('div', {
-  width: '100%',
   display: 'flex',
   justifyContent: 'space-between',
   alignItems: 'center',
   marginBottom: '$3',
   gap: '$3',
-  padding: '0 16px',
   '@sm': {
     flexDirection: 'column'
   }
@@ -46,8 +29,6 @@ const ButtonsContainer = styled('div', {
   display: 'flex',
   justifyContent: 'stretch',
   gap: '$4',
-  width: '100%',
-  padding: '0 16px',
   '@sm': {
     flexDirection: 'column',
     gap: '$3'
@@ -63,7 +44,7 @@ export const NFTDeal: FC<NFTDealProps> = observer(({
 }) => {
   useSyncAESFileKey(tokenFullId, transfer)
   return (
-    <NFTDealStyle>
+    <>
       <DealContainerInfo>
         {children}
         {order && (
@@ -78,6 +59,6 @@ export const NFTDeal: FC<NFTDealProps> = observer(({
           reFetchOrder={reFetchOrder}
         />
       </ButtonsContainer>
-    </NFTDealStyle>
+    </>
   )
 })
