+++ packages/frontend/src/app/components/NFT/NFTDeal/NFTDealActions/ActionButtons/ButtonApproveExchange.tsx
@@ -24,7 +24,6 @@ export const ButtonApproveExchange: FC<ButtonApproveExchangeProps> = ({ tokenFul
       <Button
         primary
         fullWidth
-        borderRadiusSecond
         onPress={async () => {
           await approveExchange()
           callback?.()
