+++ packages/frontend/src/app/components/NFT/NFTDeal/NFTDealActions/ActionButtons/ButtonApproveTransfer.tsx
@@ -26,7 +26,6 @@ export const ButtonApproveTransfer: FC<ButtonApproveTransferProps> = ({ tokenFul
       <Button
         primary
         fullWidth
-        borderRadiusSecond
         onPress={async () => {
           await approveTransfer()
           callback?.()
