+++ packages/frontend/src/app/components/NFT/NFTDeal/NFTDealActions/ActionButtons/ButtonCancelOrder.tsx
@@ -24,7 +24,6 @@ export const ButtonCancelOrder: FC<ButtonCancelOrderProps> = ({ tokenFullId, cal
       <Button
         primary
         fullWidth
-        borderRadiusSecond
         onPress={async () => {
           await cancelOrder()
           callback?.()
