+++ packages/frontend/src/app/components/NFT/NFTDeal/NFTDealActions/ActionButtons/ButtonFinalizeTransfer.tsx
@@ -24,7 +24,6 @@ export const ButtonFinalizeTransfer: FC<ButtonFinalizeTransferProps> = ({ tokenF
       <Button
         primary
         fullWidth
-        borderRadiusSecond
         onPress={async () => {
           await finalizeTransfer()
           callback?.()
