+++ packages/frontend/src/app/components/NFT/NFTDeal/NFTDealActions/ActionButtons/ButtonFulfillOrder.tsx
@@ -28,14 +28,13 @@ export const ButtonFulfillOrder: FC<ButtonFulfillOrderProps> = observer(({ token
       <Button
         primary
         fullWidth
-        borderRadiusSecond
         onPress={async () => {
           await fulfillOrder()
           callback?.()
         }}
         isDisabled={isLoading}
       >
-        Buy now
+        Buy
       </Button>
     </>
   )
