+++ packages/frontend/src/app/components/NFT/NFTDeal/NFTDealActions/ActionButtons/ButtonPlaceOrder.tsx
@@ -44,7 +44,6 @@ export const ButtonPlaceOrder: FC<ButtonPlaceOrderProps> = ({ tokenFullId, callb
       <Button
         primary
         fullWidth
-        borderRadiusSecond
         onPress={openModal}
         isDisabled={isLoading}
       >
