+++ packages/frontend/src/app/components/NFT/NFTDeal/NFTDealActions/ActionButtons/ButtonSetPublicKeyTransfer.tsx
@@ -24,7 +24,6 @@ export const ButtonSetPublicKeyTransfer: FC<ButtonSetPublicKeyTransferProps> = (
       <Button
         primary
         fullWidth
-        borderRadiusSecond
         onPress={async () => {
           await setPublicKey()
           callback?.()
