+++ packages/frontend/src/app/components/NFT/NFTDeal/NFTDealActions/HideAction.tsx
@@ -2,6 +2,5 @@ import { styled } from '../../../../../styles'
 import { Hide } from '../../../../UIkit/Hide/Hide'
 
 export const HideAction = styled(Hide, {
-  width: '100%',
   flexGrow: '1'
 })
