+++ packages/frontend/src/app/components/NFT/NFTDeal/NFTDealActions/NFTDealActionsBuyer.tsx
@@ -55,7 +55,6 @@ export const NFTDealActionsBuyer: FC<NFTDealActionsBuyerProps> = observer(({
             secondary
             isDisabled
             fullWidth
-            borderRadiusSecond
           >
             NFT is not listed
           </Button>
