+++ packages/frontend/src/app/components/NFT/NFTDeal/NFTDealPrice.tsx
@@ -10,23 +10,22 @@ export interface NFTDealPriceProps {
 
 export const Price = styled('div', {
   display: 'flex',
-  flexDirection: 'column',
-  width: '100%',
-  height: '64px',
-  background: 'linear-gradient(0deg, rgba(255, 255, 255, 0.9), rgba(255, 255, 255, 0.9)), #232528',
-  borderRadius: '16px',
-  padding: '8px 16px'
+  gap: '$1',
+  flexDirection: 'column'
 })
 
 export const PriceTitle = styled('span', {
-  ...textVariant('primary2').true,
+  ...textVariant('primary1'),
   color: '$gray500'
 })
 
 export const PriceValue = styled('p', {
-  fontSize: '24px',
-  fontWeight: 600,
-  color: '$gray800'
+  fontSize: '$h3',
+  fontWeight: 700,
+  color: '$blue900',
+  '@xl': {
+    fontSize: '$h4'
+  }
 })
 
 export const NFTDealPrice: FC<NFTDealPriceProps> = observer(({ price }) => {
