+++ packages/frontend/src/styles/stitches.config.ts
@@ -137,7 +137,6 @@ export const {
 
     breakpoints: {
       sm: '600px',
-      mdx: '700px',
       md: '900px',
       lg: '1200px',
       xl: '1536px'
@@ -156,13 +155,11 @@ export const {
 
   media: {
     sm: '(max-width: 600px)',
-    mdx: '(max-width: 700px)',
     md: '(max-width: 900px)',
     lg: '(max-width: 1200px)',
     xl: '(max-width: 1536px)',
 
     smUp: '(min-width: 600px)',
-    mdxUp: '(min-width: 700px)',
     mdUp: '(min-width: 900px)',
     lgUp: '(min-width: 1200px)',
     xlUp: '(min-width: 1536px)'
